Fix all CI failures: Clippy warnings, TS scripts, Mojo CI config
- Fix 257+ Clippy warnings across Nucleus (56 source files) via auto-fix and manual fixes
- Fix 25+ Clippy warnings across RS framework (19 source files)
- Add missing "typecheck" script to ts/package.json for CI
- Fix Mojo CI workflow to use pixi manifest-path for non-root project

All 2187 Nucleus tests and 800 RS tests passing.

+++ rs/crates/neutron-redis/src/session.rs
@@ -110,7 +110,7 @@ impl SessionStore for RedisSessionStore {
     ) {
         let key          = self.redis_key(id);
         let mut conn     = self.pool.conn();
-        let ttl_secs     = ttl.as_secs().max(1) as u64;
+        let ttl_secs     = ttl.as_secs().max(1);
 
         let serialised = match serde_json::to_string(&data) {
             Ok(s)  => s,
